backend/requirements.txt
# Core Flask and API
Flask==2.3.3
Flask-CORS==4.0.0

# Database connectivity
pyodbc==4.0.39

# Data processing and analysis
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1

# Scientific computing for advanced calculations
scipy==1.11.2

# Utilities
Werkzeug==2.3.7
python-dotenv==1.0.0

# Development and testing
pytest==7.4.2
pytest-cov==4.1.0